Conflicts:
	BDT-JPA/src/cl/minsal/divap/model/DocumentoEstimacionflujocaja.java
	BDT-Rest/src/minsal/divap/rest/EstimacionFlujoCajaRESTService.java
	BDT-ejb/ejbModule/minsal/divap/dao/DocumentDAO.java
	BDT-ejb/ejbModule/minsal/divap/excel/GeneradorExcel.java
	BDT-ejb/ejbModule/minsal/divap/service/AlfrescoService.java
	BDT-ejb/ejbModule/minsal/divap/service/EstimacionFlujoCajaService.java

merge con repo, incluidos cambios de pacheco

+++ BDT-ejb/ejbModule/minsal/divap/vo/SubtituloFlujoCajaVO.java
@@ -18,11 +18,11 @@ public class SubtituloFlujoCajaVO implements Serializable {
 	private Integer idSubtitulo;
 	private String subtitulo;
 	private Integer idMarcoPresupuestario;
-	private Integer marcoPresupuestario;
+	private Long marcoPresupuestario;
 	private TransferenciaSummaryVO transferenciaAcumulada;
 	private ConveniosSummaryVO convenioRecibido;
 	private List<CajaMontoSummaryVO> cajaMontos;
-	private Integer totalMontos;
+	private Long totalMontos;
 	private String color = "#FFB5B5";
 
 	public SubtituloFlujoCajaVO() {
@@ -35,10 +35,10 @@ public class SubtituloFlujoCajaVO implements Serializable {
 	
 	public SubtituloFlujoCajaVO(Integer idServicio, String servicio,
 			Integer idSubtitulo, String subtitulo,
-			Integer idMarcoPresupuestario, Integer marcoPresupuestario,
+			Integer idMarcoPresupuestario, Long marcoPresupuestario,
 			TransferenciaSummaryVO transferenciaAcumulada,
 			ConveniosSummaryVO convenioRecibido,
-			List<CajaMontoSummaryVO> cajaMontos, Integer totalMontos,
+			List<CajaMontoSummaryVO> cajaMontos, Long totalMontos,
 			String color) {
 		super();
 		this.idServicio = idServicio;
@@ -55,11 +55,11 @@ public class SubtituloFlujoCajaVO implements Serializable {
 	}
 
 
-	public Integer getMarcoPresupuestario() {
+	public Long getMarcoPresupuestario() {
 		return marcoPresupuestario;
 	}
 
-	public void setMarcoPresupuestario(Integer marcoPresupuestario) {
+	public void setMarcoPresupuestario(Long marcoPresupuestario) {
 		this.marcoPresupuestario = marcoPresupuestario;
 	}
 
@@ -112,8 +112,8 @@ public class SubtituloFlujoCajaVO implements Serializable {
 		this.cajaMontos = cajaMontos;
 	}
 
-	public Integer getTotalMontos() {
-		totalMontos = 0;
+	public Long getTotalMontos() {
+		totalMontos = 0L;
 		if(cajaMontos != null && cajaMontos.size() > 0){
 			for(CajaMontoSummaryVO monto : cajaMontos){
 				totalMontos += monto.getMontoMes();
@@ -122,7 +122,7 @@ public class SubtituloFlujoCajaVO implements Serializable {
 		return totalMontos;
 	}
 
-	public void setTotalMontos(Integer totalMontos) {
+	public void setTotalMontos(Long totalMontos) {
 		this.totalMontos = totalMontos;
 	}
